Move distance snap toggle button implementation to `DistancedHitObjectComposer`
