+++ osu.Game.Rulesets.Osu/Edit/OsuHitObjectComposer.cs
@@ -46,12 +46,10 @@ namespace osu.Game.Rulesets.Osu.Edit
             new SpinnerCompositionTool()
         };
 
-        private readonly Bindable<TernaryState> distanceSnapToggle = new Bindable<TernaryState>();
         private readonly Bindable<TernaryState> rectangularGridSnapToggle = new Bindable<TernaryState>();
 
         protected override IEnumerable<TernaryButton> CreateTernaryButtons() => base.CreateTernaryButtons().Concat(new[]
         {
-            new TernaryButton(distanceSnapToggle, "Distance Snap", () => new SpriteIcon { Icon = FontAwesome.Solid.Ruler }),
             new TernaryButton(rectangularGridSnapToggle, "Grid Snap", () => new SpriteIcon { Icon = FontAwesome.Solid.Th })
         });
 
@@ -82,7 +80,7 @@ namespace osu.Game.Rulesets.Osu.Edit
 
             placementObject = EditorBeatmap.PlacementObject.GetBoundCopy();
             placementObject.ValueChanged += _ => updateDistanceSnapGrid();
-            distanceSnapToggle.ValueChanged += _ => updateDistanceSnapGrid();
+            DistanceSnapToggle.ValueChanged += _ => updateDistanceSnapGrid();
 
             // we may be entering the screen with a selection already active
             updateDistanceSnapGrid();
@@ -128,7 +126,7 @@ namespace osu.Game.Rulesets.Osu.Edit
 
             if (snapType.HasFlagFast(SnapType.Grids))
             {
-                if (distanceSnapToggle.Value == TernaryState.True && distanceSnapGrid != null)
+                if (DistanceSnapToggle.Value == TernaryState.True && distanceSnapGrid != null)
                 {
                     (Vector2 pos, double time) = distanceSnapGrid.GetSnappedPosition(distanceSnapGrid.ToLocalSpace(screenSpacePosition));
 
@@ -197,7 +195,7 @@ namespace osu.Game.Rulesets.Osu.Edit
             distanceSnapGridCache.Invalidate();
             distanceSnapGrid = null;
 
-            if (distanceSnapToggle.Value != TernaryState.True)
+            if (DistanceSnapToggle.Value != TernaryState.True)
                 return;
 
             switch (BlueprintContainer.CurrentTool)
@@ -242,24 +240,15 @@ namespace osu.Game.Rulesets.Osu.Edit
         protected override bool AdjustDistanceSpacing(GlobalAction action, float amount)
         {
             // To allow better visualisation, ensure that the spacing grid is visible before adjusting.
-            distanceSnapToggle.Value = TernaryState.True;
+            DistanceSnapToggle.Value = TernaryState.True;
 
             return base.AdjustDistanceSpacing(action, amount);
         }
 
-        private bool distanceSnapMomentary;
         private bool gridSnapMomentary;
 
         private void handleToggleViaKey(KeyboardEvent key)
         {
-            bool altPressed = key.AltPressed;
-
-            if (altPressed != distanceSnapMomentary)
-            {
-                distanceSnapMomentary = altPressed;
-                distanceSnapToggle.Value = distanceSnapToggle.Value == TernaryState.False ? TernaryState.True : TernaryState.False;
-            }
-
             bool shiftPressed = key.ShiftPressed;
 
             if (shiftPressed != gridSnapMomentary)
